requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
gspread==6.0.0
google-auth==2.23.4
google-auth-oauthlib==1.0.0
matplotlib==3.8.0
seaborn==0.13.2
python-dateutil==2.8.2
streamlit-calendar==1.0.0
scikit-learn==1.3.2
